refactor api module architecture
