+++ application-api/src/main/java/org/arispay/controller/ValidationController.java
@@ -1,7 +1,7 @@
 package org.arispay.controller;
 
 import org.arispay.data.ClientDto;
-import org.arispay.data.dtorequest.validation.ValidationRequest;
+import org.arispay.data.fbl.dtorequest.validation.ValidationRequest;
 import org.arispay.data.dtoresponse.validation.ValidationResponse;
 import org.arispay.ports.api.ClientServicePort;
 import org.springframework.beans.factory.annotation.Autowired;
