+++ application-api/src/main/java/org/arispay/controller/fbl/CollectionsController.java
@@ -1,0 +1,7 @@
+package org.arispay.controller.fbl;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class CollectionsController {
+}
