+++ domain/src/main/java/org/arispay/data/GenericHttpResponse.java
@@ -1,4 +1,4 @@
-package org.arispay.data.dtoresponse;
+package org.arispay.data;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,7 +8,7 @@ import org.springframework.http.HttpStatus;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class ErrorResp {
+public class GenericHttpResponse {
 	HttpStatus httpStatus;
 	String message;
 }
