+++ domain/src/main/java/org/arispay/data/dtoauth/JwtLoginReq.java
@@ -1,4 +1,4 @@
-package org.arispay.data.dtorequest;
+package org.arispay.data.dtoauth;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
