+++ domain/src/main/java/org/arispay/data/fbl/dtorequest/validation/ValidationRequest.java
@@ -1,4 +1,4 @@
-package org.arispay.data.dtorequest.confirmation;
+package org.arispay.data.fbl.dtorequest.validation;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,8 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ConfirmationRequest {
+public class ValidationRequest {
+
     private String action;
     private Payload payload;
 }
